+++ .gitignore
@@ -5,3 +5,5 @@
 /node_modules
 /umd
 npm-debug.log*
+yarn.*
+.idea/
