+++ tests/index-test.js
@@ -2,7 +2,9 @@ import expect from 'expect'
 import React from 'react'
 import {render, unmountComponentAtNode} from 'react-dom'
 
-import Component from 'src/'
+import CoinHiveClient, { Shortlink } from 'src/index'
+
+const SECRET = '';
 
 describe('Component', () => {
   let node
@@ -16,8 +18,17 @@ describe('Component', () => {
   })
 
   it('displays a welcome message', () => {
-    render(<Component/>, node, () => {
+    render(<CoinHiveClient/>, node, () => {
       expect(node.innerHTML).toContain('Welcome to React components')
     })
   })
 })
+
+describe('Shortlink', () => {
+    it('returns a link', () => {
+        const linker = new Shortlink(SECRET, true);
+        linker('http://mcordes.me/')
+            .then(res => expect(JSON.stringify(res)).toContain("success"))
+            .catch(err => expect(false).toBe(true));
+    });
+});
